User/requirement.txt
flask-mysqldb
flask-mysql
sudo apt-get install libmysqlclient-dev
